Bump zope.container to 4.6 in tests with Python 3.8:
With zope.container==4.5 (which is the version pinned by the Plone 5.2 KGS)
we get test failures like these with Python 3.8:

```
  File ".../zope.container-4.5.0-py3.8-linux-x86_64.egg/zope/container/_proxy.py", line 58, in __setstate__
    object.__setattr__(self, '__parent__', state[0])
TypeError: can't apply this __setattr__ to ContainedProxy object
```

(when accessing portlets storage in the `test_copy_dashboard_no_replace` test).

This could possibly be related to https://bugs.python.org/issue41295, but
that issue was supposedly resolved in Python 3.8.5 - yet the problem still
occurs with Python 3.6.18 (on macOS as well as Linux).

It appears that the changes from zopefoundation/zope.container#44 finally
fix this issue, so I'm bumping zope.container to 4.6 in tests for now.